Added SPR function as computed in Peter's reports

+++ src/main/java/uk/ac/ox/oxfish/biology/complicated/CaliforniaStockAssessmentGrowthBinParameters.java
@@ -159,7 +159,7 @@ public class CaliforniaStockAssessmentGrowthBinParameters implements Meristics{
      * @return the length of the fish
      */
     @Override
-    public double getLengthAtAge(double ageInYears, int subdivision) {
+    public double getLengthAtAge(int ageInYears, int subdivision) {
 
         return getLength(subdivision,ageInYears > maxAge ? (int)ageInYears : maxAge );
     }
