+++ src/main/java/uk/ac/ox/oxfish/biology/complicated/FromListMeristics.java
@@ -104,7 +104,7 @@ public class FromListMeristics implements Meristics {
      * @return the length of the fish
      */
     @Override
-    public double getLengthAtAge(double ageInYears, int subdivision) {
+    public double getLengthAtAge(int ageInYears, int subdivision) {
         return growth.getLengthAtAge(ageInYears,subdivision);
     }
 }
