+++ src/main/java/uk/ac/ox/oxfish/biology/complicated/Meristics.java
@@ -20,8 +20,6 @@
 
 package uk.ac.ox.oxfish.biology.complicated;
 
-import com.google.common.collect.ImmutableList;
-
 /**
  * this interface keeps track of basically how much a fish in a bin weigh and is long as well as how many bins and
  * subdivisions there actually are
@@ -55,6 +53,6 @@ public interface Meristics {
      * @param subdivision the subdivision we are study (male/female is different for example)
      * @return the length of the fish
      */
-    double getLengthAtAge(double ageInYears, int subdivision);
+    double getLengthAtAge(int ageInYears, int subdivision);
 
 }
